src/requirements.txt
cloudshell-automation-api>=7.1.0.0,<7.2.0.0
cloudshell-shell-core>=2.3.0,<2.4.0